eliminadas interfaces globales
