+++ pages/admin/trabajadores/index.tsx
@@ -1,9 +1,14 @@
 import type { NextPage } from 'next'
 import TrabajadorRow from '../../../components/TrabajadorRow';
 import { Table } from '@mantine/core'
-import { TrabajadorRowProps } from '../../../interfaces';
 import Head from 'next/head';
 
+export interface TrabajadorRowProps {
+  Name: string;
+  Rol: string;
+  Last_access: string; 
+  Foto: string;
+}
 
 const elements: TrabajadorRowProps[] = [
   {
@@ -28,7 +33,7 @@ const elements: TrabajadorRowProps[] = [
     Name: "Eduardo Pinto",
     Rol: "Worker",
     Last_access: 'Connected',
-    Foto: {}
+    Foto: "https://d2qp0siotla746.cloudfront.net/img/use-cases/profile-picture/template_2.jpg"
   }
 ];
